# api_docs/api-docs-input.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest
      summary: Returns a specific user
      operationId: profile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User data returned correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
    put:
      tags:
      - user-rest
      summary: Edits a user profile
      operationId: editProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "403":
          description: Operation not allowed for the current user
          content:
            application/json: {}
        "400":
          description: Exception originated from JSON data processing or mapping
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "200":
          description: User profile edited correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
    delete:
      tags:
      - user-rest
      summary: Deletes a user profile
      operationId: deleteProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Operation successful
        "403":
          description: Operation only allowed to admins or the user itself
          content:
            application/json: {}
        "404":
          description: Profile not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE'
...
  /api/reviews/{reviewId}/users/{userId}:
    put:
      tags:
      - review-rest
      summary: Edit a review
      description: Edit a review by review ID and user ID.
      operationId: editReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "404":
          description: User or review not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "200":
          description: Review edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
...
  /api/users:
    get:
      tags:
      - user-rest
      summary: Returns a list of all users
      operationId: loadAllUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "403":
          description: Operation only allowed for admin users
          content:
            application/json: {}
        "404":
          description: Users not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseUserE_HotelDetails'
        "200":
          description: List of users returned correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_HotelDetails'
    post:
      tags:
      - user-rest
      summary: Registers a new user
      operationId: registerClient
      parameters:
      - name: type
        in: query
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserE'
        required: true
      responses:
        "409":
          description: User already exists
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE'
        "400":
          description: "Bad request, maybe one of the user attributes is missing or\
            \ the type is not valid"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE'
        "201":
          description: User registered correctly
  /api/users/{id}/image:
    get:
      tags:
      - user-image-rest
      summary: Download user profile image
      operationId: downloadProfileImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User profile image downloaded successfully
          content:
            image/*: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - user-image-rest
      summary: Upload user profile image
      operationId: uploadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "403":
          description: Operation only allowed for the user or an admin
          content:
            application/json: {}
        "201":
          description: User profile image uploaded successfully
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - user-image-rest
      summary: Delete user profile image
      operationId: deleteImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Operation only allowed for the user or an admin
          content:
            application/json: {}
        "204":
          description: User profile image deleted successfully
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
...
  /api/hotels/{id}/image:
    get:
      tags:
      - hotel-image-rest
      summary: Download hotel image
      description: Download the image of a specific hotel by hotel ID.
      operationId: downloadProfileImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Image retrieved successfully
          content:
            image/*: {}
    post:
      tags:
      - hotel-image-rest
      summary: Upload hotel image
      description: Upload an image for a specific hotel by hotel ID.
      operationId: uploadImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json: {}
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - hotel-image-rest
      summary: Delete hotel image
      description: Delete the image of a specific hotel by hotel ID.
      operationId: deleteImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Image deleted successfully
          content:
            application/json: {}
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
...
  /api/users/{id}/type/user:
    get:
      tags:
      - user-rest
      summary: Returns if the user is a user or not
      operationId: isUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Boolean matching user role returned correctly
          content:
            application/json:
              schema:
                type: boolean
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: boolean
...
  /api/reviews/users/{id}:
    get:
      tags:
      - review-rest
      summary: Get user reviews
      description: Get all reviews of a specific user by user ID.
      operationId: userReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReviewDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReview_ReviewDetails'
...
  /api/reservations/users/{id}:
    get:
      tags:
      - reservation-rest
      summary: Get user reservations
      description: Returns a page of reservations for a specific user.
      operationId: getUserReservations
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reservations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReservationDetails'
        "403":
          description: "Forbidden, the user does not have permission to view these\
            \ reservations"
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReservation_ReservationDetails'
...
components:
  schemas:
    Hotel_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        rooms:
          $ref: '#/components/schemas/Room_UserDetails'
        idReservation:
          type: integer
          format: int64
    Review_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        idReview:
          type: integer
          format: int64
    Room_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
    UserE_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_UserDetails'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review_UserDetails'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/Hotel_UserDetails'
        imageBase64:
          type: string
...
    Hotel_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Review_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel_ReviewDetails'
        user:
          $ref: '#/components/schemas/UserE_ReviewDetails'
        idReview:
          type: integer
          format: int64
    UserE_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        imageBase64:
          type: string
    Hotel_ReservationDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation_ReservationDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        hotel:
          $ref: '#/components/schemas/Hotel_ReservationDetails'
        room:
          $ref: '#/components/schemas/Room_ReservationDetails'
        user:
          $ref: '#/components/schemas/UserE_ReservationDetails'
        rooms:
          $ref: '#/components/schemas/Room_ReservationDetails'
        idReservation:
          type: integer
          format: int64
    Room_ReservationDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
    UserE_ReservationDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        imageBase64:
          type: string
...
    UserE_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        imageBase64:
          type: string
    Hotel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        manager:
          $ref: '#/components/schemas/UserE'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        hotel:
          $ref: '#/components/schemas/Hotel'
        room:
          $ref: '#/components/schemas/Room'
        user:
          $ref: '#/components/schemas/UserE'
        rooms:
          $ref: '#/components/schemas/Room'
        idReservation:
          type: integer
          format: int64
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel'
        user:
          $ref: '#/components/schemas/UserE'
        idReview:
          type: integer
          format: int64
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        hotel:
          $ref: '#/components/schemas/Hotel'
    UserE:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/Hotel'
        imageBase64:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageResponse_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseUserE_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserE_HotelDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponse_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseReview_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review_ReviewDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponse_ReservationDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseReservation_ReservationDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_ReservationDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...